my_agent/app/requirement.txt
google-adk
google-genai
fastapi
uvicorn[standard]
pydantic
pypdf
